+++ libs/components/list-builder/projects/list-builder/src/modules/list/state/items/set-items-selected.action.ts
@@ -1,0 +1,11 @@
+/**
+ * @internal
+ */
+export class ListItemsSetSelectedAction {
+  constructor(
+    public items: string[],
+    /* istanbul ignore next */
+    public selected: boolean = false,
+    public refresh: boolean = true
+  ) {}
+}