+++ libs/components/list-builder/src/app/docs/list-filters/demo/list-filters-docs-modal-context.ts
@@ -1,3 +1,0 @@
-export class ListFiltersDocsModalContext {
-  public appliedFilters: any[];
-}