+++ libs/components/list-builder/src/app/docs/list/list-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-list-docs',
-  templateUrl: './list-docs.component.html'
-})
-export class ListDocsComponent { }